better structure and more fonts
